requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
gunicorn==21.2.0
mistralai==0.4.0
tavily-python==0.3.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiohttp==3.9.3
aiofiles==23.1.0
aiologger==0.7.0
feedparser==6.0.10